// api/models/callingType.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * MagicMessenger
 * OpenAPI spec version: v1
 */

export type CallingType = (typeof CallingType)[keyof typeof CallingType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CallingType = {
  Video: "Video",
  Audio: "Audio",
} as const;